Reorganise pagination
